website/backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
requests==2.31.0
openai==1.3.9
google-generativeai==0.3.0
google-auth==2.28.2

# ML dependencies for playground
joblib==1.3.2
scikit-learn==1.4.0
xgboost==2.0.3
yfinance==0.2.36
ta==0.11.0
pandas==2.2.0
numpy==1.26.3
